:zap:  restructure to trip to component
