+++ web/src/components/Trip.module.css
@@ -1,0 +1,126 @@
+
+
+.heading1{ 
+    font-size: 45px;
+    font-family:Helvethaica bold;
+    display: inline;
+    
+}
+.heading2{ 
+    font-size: 75px;
+    font-family:Helvethaica bold;
+    color: #2C9CDB;
+    margin-top: 30px;
+
+}
+.heading3{ 
+    font-size: 25px;
+    font-family:Helvethaica;
+
+}
+.topic{
+    font-size:20px;
+    font-family:Helvethaica;
+}
+
+
+
+.row{
+    display: flex;
+    flex-direction: row;
+   
+   
+}
+
+.img{
+    width: 40%;
+    height:100%;
+    border-radius: 20px;
+    margin-right: 30px;
+    
+}
+
+.detail{
+    height: 100%;
+    width: 100%;
+
+    border-radius: 20px;
+    
+}
+
+.column{
+    display: flex;
+    flex-direction: column;
+}
+
+.smallImage{
+    width:150px;
+    height:150px;
+   
+    margin-Right:30px;
+    border-radius: 20px;
+}
+
+.topic{
+    font-size: 35px;
+}
+.content{ 
+    font-size: 20px;
+}
+
+.mainImg{
+    width: 100%;
+    height: 100%;
+    object-fit:cover;
+    border-radius: 20px;
+    
+  
+}
+
+.tags{
+   
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+    /* white-space: nowrap; */
+}
+
+.tripRow{
+    display: flex;
+    flex-direction: row;
+  
+    width: 80%;
+    height: 500px;
+    margin-top:30px;
+   
+}
+@media (max-width: 1000px) {
+    
+    .img{ 
+        display: none;
+    }
+   }
+   @media (max-width: 600px) {
+    
+    .mainImg{
+        display: none;
+        
+      
+   }
+   .tags{
+   
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+      /* white-space: nowrap; */
+  }
+  .tripRow{
+   
+   margin-bottom: 100px;
+   
+}
+  
+}
+
+
+
